+++ forest/benchmarking/tests/test_observable_estimation.py
@@ -2,7 +2,6 @@ import functools
 import itertools
 import random
 random.seed(1)  # seed random number generation for all calls to rand_ops
-from math import pi
 import numpy as np
 from operator import mul
 
@@ -14,8 +13,7 @@ from pyquil.api import WavefunctionSimulator, QVMConnection
 from pyquil.paulis import sI, sX, sY, sZ, PauliSum, PauliTerm
 from forest.benchmarking.observable_estimation import *
 from forest.benchmarking.observable_estimation import _OneQState,\
-    _max_tpb_overlap, _max_weight_operator, _max_weight_state, \
-    _flip_array_to_prog,  _measure_bitstrings, _one_q_sic_prep
+    _max_tpb_overlap, _max_weight_operator, _max_weight_state, _one_q_sic_prep
 
 
 def _generate_random_states(n_qubits, n_terms):
@@ -480,7 +478,7 @@ def test_estimate_observables_symmetrize(forest):
 
     qc = get_qc('2q-qvm')
     qc.qam.random_seed = 1
-    for res in estimate_observables(qc, gsuite, symmetrization_method=exhaustive_symmetrization):
+    for res in estimate_observables(qc, gsuite, symm_type=-1):
         if res.setting.observable in [sI(), sZ(0), sZ(1), sZ(0) * sZ(1)]:
             assert np.abs(res.expectation) > 0.9
         else:
@@ -503,7 +501,7 @@ def test_estimate_observables_symmetrize_calibrate(forest):
     qc = get_qc('2q-qvm')
     qc.qam.random_seed = 1
     results = list(estimate_observables(qc, gsuite,
-                                        symmetrization_method=exhaustive_symmetrization))
+                                        symm_type=-1))
     for res in calibrate_observable_estimates(qc, results):
         if res.setting.observable in [sI(), sZ(0), sZ(1), sZ(0) * sZ(1)]:
             assert np.abs(res.expectation) > 0.9
@@ -522,19 +520,6 @@ def test_estimate_observables_zero_expectation(forest):
                                  program=Program(I(0)))
     result = list(estimate_observables(qc, suite))[0]
     np.testing.assert_almost_equal(result.expectation, 0.0, decimal=1)
-
-
-def test_flip_array_to_prog():
-    qubits = [0, 2, 3]
-    ops_strings = list(itertools.product([0, 1], repeat=len(qubits)))
-    d_expected = {(0, 0, 0): '', (0, 0, 1): 'RX(pi) 3\n', (0, 1, 0): 'RX(pi) 2\n',
-                  (0, 1, 1): 'RX(pi) 2\nRX(pi) 3\n',
-                  (1, 0, 0): 'RX(pi) 0\n', (1, 0, 1): 'RX(pi) 0\nRX(pi) 3\n',
-                  (1, 1, 0): 'RX(pi) 0\nRX(pi) 2\n',
-                  (1, 1, 1): 'RX(pi) 0\nRX(pi) 2\nRX(pi) 3\n'}
-    for op_str in ops_strings:
-        p = _flip_array_to_prog(op_str, qubits)
-        assert str(p) == d_expected[op_str]
 
 
 def test_shots_to_obs_moments():
@@ -613,7 +598,7 @@ def test_estimate_observables_uncalibrated_symmetric_readout(forest):
     uncalibr_e = np.zeros(runs * len(expt_list))
 
     for idx, res in enumerate(estimate_observables(qc, obs_expt, num_shots=1000,
-                                                   symmetrization_method=exhaustive_symmetrization)):
+                                                   symm_type=-1)):
         uncalibr_e[idx] = res.expectation
 
     assert np.isclose(np.mean(uncalibr_e[::3]), expected_expectation_z_basis, atol=3e-2)
@@ -637,7 +622,7 @@ def test_estimate_observables_calibrated_symmetric_readout(forest):
     for sim_num in range(num_simulations):
         qc.qam.random_seed = sim_num+1
         expt_results = list(estimate_observables(qc, obs_expt, num_shots=1000,
-                                                 symmetrization_method=exhaustive_symmetrization))
+                                                 symm_type=-1))
         expt_results = list(calibrate_observable_estimates(qc, expt_results, noisy_program=p))
         expectations.append([res.expectation for res in expt_results])
     expectations = np.array(expectations)
@@ -664,7 +649,7 @@ def test_estimate_observables_result_zero_symmetrization_calibration(forest):
     for sim_num in range(num_simulations):
         qc.qam.random_seed = sim_num+1
         expt_results = list(estimate_observables(qc, obs_expt, num_shots=1000,
-                                                 symmetrization_method=exhaustive_symmetrization))
+                                                 symm_type=-1))
         expt_results = list(calibrate_observable_estimates(qc, expt_results, noisy_program=p))
         expectations.append([res.expectation for res in expt_results])
         raw_expectations.append([res.raw_expectation for res in expt_results])
@@ -742,7 +727,7 @@ def test_estimate_observables_2q_readout_error_one_measured(forest):
     cal_e = np.zeros(runs)
 
     results = calibrate_observable_estimates(qc, list(estimate_observables(qc, expt, num_shots=500,
-                                                  symmetrization_method=exhaustive_symmetrization)),
+                                                  symm_type=-1)),
                                              num_shots=500*runs, noisy_program=p)
     for idx, res in enumerate(results):
         raw_e[idx] = res.raw_expectation
@@ -752,61 +737,6 @@ def test_estimate_observables_2q_readout_error_one_measured(forest):
     assert np.isclose(np.mean(raw_e), 0.849, atol=3e-2)
     assert np.isclose(np.mean(obs_e), 1.0, atol=3e-2)
     assert np.isclose(np.mean(cal_e), 0.849, atol=3e-2)
-
-
-def test_exhaustive_symmetrization_1q(forest):
-    qc = get_qc('9q-qvm')
-    qc.qam.random_seed = 1
-    qubs = [5]
-    num_shots = 1000
-    p = Program()
-    p00, p11 = 0.90, 0.80
-    p.define_noisy_readout(5, p00, p11)
-
-    symm_programs, symm_prog_qs, flip_arrays, prog_groups = exhaustive_symmetrization([p], [qubs])
-    symm_outputs = _measure_bitstrings(qc, symm_programs, symm_prog_qs, num_shots=num_shots)
-    results = consolidate_symmetrization_outputs(symm_outputs, flip_arrays, prog_groups)[0]
-
-    total_num_shots = 2**len(qubs) * num_shots
-
-    assert total_num_shots == len(results)
-
-    frac0 = np.count_nonzero(results == 0) / total_num_shots
-    expected_frac0 = (p00 + p11) / 2
-
-    assert symm_prog_qs == [[5]] * 2
-    assert np.isclose(frac0, expected_frac0, atol=3e-2)
-
-
-def test_exhaustive_symmetrization_2q(forest):
-    qc = get_qc('9q-qvm')
-    qc.qam.random_seed = 1
-    qubs = [5, 7]
-    num_shots = 1000
-    p = Program()
-    p5_00, p5_11 = 0.90, 0.80
-    p7_00, p7_11 = 0.99, 0.77
-    p.define_noisy_readout(5, p5_00, p5_11)
-    p.define_noisy_readout(7, p7_00, p7_11)
-
-    symm_programs, symm_prog_qs, flip_arrays, prog_groups = exhaustive_symmetrization([p], [qubs])
-    symm_outputs = _measure_bitstrings(qc, symm_programs, symm_prog_qs, num_shots=num_shots)
-    results = consolidate_symmetrization_outputs(symm_outputs, flip_arrays, prog_groups)[0]
-
-    assert symm_prog_qs == [[5,7]]*4
-
-    total_num_shots = 2**len(qubs) * num_shots
-
-    assert total_num_shots == len(results)
-
-    frac5_0 = np.count_nonzero(results[:, 0] == 0) / total_num_shots
-    frac7_0 = np.count_nonzero(results[:, 1] == 0) / total_num_shots
-
-    expected_frac5_0 = (p5_00 + p5_11) / 2
-    expected_frac7_0 = (p7_00 + p7_11) / 2
-
-    assert np.isclose(frac5_0, expected_frac5_0, atol=3e-2)
-    assert np.isclose(frac7_0, expected_frac7_0, atol=3e-2)
 
 
 def test_estimate_observables_inherit_noise_errors(forest):
@@ -1169,7 +1099,7 @@ def test_bit_flip_channel_fidelity_readout_error(forest):
         qc.qam.random_seed = sim_num+1
         expt_results = []
         results = estimate_observables(qc, process_exp,
-                                       symmetrization_method=exhaustive_symmetrization)
+                                       symm_type=-1)
         for res in calibrate_observable_estimates(qc, list(results), noisy_program=p):
             expt_results.append(res.expectation)
         expts.append(expt_results)
@@ -1214,7 +1144,7 @@ def test_depolarizing_channel_fidelity_readout_error(forest):
         qc.qam.random_seed = sim_num+1
         expt_results = []
         results = estimate_observables(qc, process_exp,
-                                       symmetrization_method=exhaustive_symmetrization)
+                                       symm_type=-1)
         for res in calibrate_observable_estimates(qc, list(results), noisy_program=p):
             expt_results.append(res.expectation)
         expts.append(expt_results)
@@ -1253,7 +1183,7 @@ def test_unitary_channel_fidelity_readout_error(forest):
         qc.qam.random_seed = sim_num+1
         expt_results = []
         results = estimate_observables(qc, process_exp,
-                                       symmetrization_method=exhaustive_symmetrization)
+                                       symm_type=-1)
         for res in calibrate_observable_estimates(qc, list(results), noisy_program=p):
             expt_results.append(res.expectation)
         expts.append(expt_results)
@@ -1314,7 +1244,7 @@ def test_2q_unitary_channel_fidelity_readout_error(forest):
         qc.qam.random_seed = sim_num+1
         expt_results = []
         results = estimate_observables(qc, process_exp,
-                                       symmetrization_method=exhaustive_symmetrization)
+                                       symm_type=-1)
         for res in calibrate_observable_estimates(qc, list(results), noisy_program=p):
             expt_results.append(res.expectation)
         expts.append(expt_results)
@@ -1342,7 +1272,7 @@ def test_measure_1q_observable_raw_expectation(forest):
     for sim_num in range(num_simulations):
         qc.qam.random_seed = sim_num+1
         expt_results = list(estimate_observables(qc, obs_expt, num_shots=1000,
-                                                 symmetrization_method=exhaustive_symmetrization))
+                                                 symm_type=-1))
         expt_results = list(calibrate_observable_estimates(qc, expt_results, noisy_program=p))
         raw_expectations.append([res.raw_expectation for res in expt_results])
     raw_expectations = np.array(raw_expectations)
@@ -1371,7 +1301,7 @@ def test_measure_1q_observable_raw_variance(forest):
     for sim_num in range(num_simulations):
         qc.qam.random_seed = sim_num+1
         expt_results = list(estimate_observables(qc, obs_expt, num_shots=num_shots,
-                                                 symmetrization_method=exhaustive_symmetrization))
+                                                 symm_type=-1))
         expt_results = list(calibrate_observable_estimates(qc, expt_results, noisy_program=p))
         raw_std_errs.append([res.raw_std_err for res in expt_results])
     raw_std_errs = np.array(raw_std_errs)
@@ -1399,7 +1329,7 @@ def test_measure_1q_observable_calibration_expectation(forest):
     for sim_num in range(num_simulations):
         qc.qam.random_seed = sim_num+1
         expt_results = list(estimate_observables(qc, obs_expt, num_shots=1000,
-                                                 symmetrization_method=exhaustive_symmetrization))
+                                                 symm_type=-1))
         expt_results = list(calibrate_observable_estimates(qc, expt_results, noisy_program=p))
         calibration_expectations.append([res.calibration_expectation for res in expt_results])
     calibration_expectations = np.array(calibration_expectations)
@@ -1428,7 +1358,7 @@ def test_measure_1q_observable_calibration_variance(forest):
     for sim_num in range(num_simulations):
         qc.qam.random_seed = sim_num+1
         expt_results = list(estimate_observables(qc, obs_expt, num_shots=num_shots,
-                                                 symmetrization_method=exhaustive_symmetrization))
+                                                 symm_type=-1))
         expt_results = list(calibrate_observable_estimates(qc, expt_results, noisy_program=p))
         raw_std_errs.append([res.raw_std_err for res in expt_results])
     raw_std_errs = np.array(raw_std_errs)
@@ -1492,7 +1422,7 @@ def test_uncalibrated_symmetric_readout_nontrivial_1q_state(forest):
 
     for idx, res in enumerate(estimate_observables(qc,
                                                   obs_expt, num_shots=1000,
-                                                  symmetrization_method=exhaustive_symmetrization)):
+                                                  symm_type=-1)):
         expect_arr[idx] = res.expectation
 
     assert np.isclose(np.mean(expect_arr), expected_expectation, atol=3e-2)
@@ -1521,7 +1451,7 @@ def test_calibrated_symmetric_readout_nontrivial_1q_state(forest):
 
     results = calibrate_observable_estimates(qc, list(estimate_observables(qc,
                                                   obs_expt, num_shots=750,
-                                                  symmetrization_method=exhaustive_symmetrization)),
+                                                  symm_type=-1)),
                                              num_shots=750, noisy_program=p)
     for idx, res in enumerate(results):
         expect_arr[idx] = res.expectation
@@ -1554,7 +1484,7 @@ def test_measure_2q_observable_raw_statistics(forest):
     for sim_num in range(num_simulations):
         qc.qam.random_seed = sim_num+1
         expt_results = list(estimate_observables(qc, obs_expt, num_shots=num_shots,
-                                                 symmetrization_method=exhaustive_symmetrization))
+                                                 symm_type=-1))
         expt_results = list(calibrate_observable_estimates(qc, expt_results, noisy_program=p))
         raw_expectations.append([res.raw_expectation for res in expt_results])
         raw_std_errs.append([res.raw_std_err for res in expt_results])
@@ -1602,7 +1532,7 @@ def test_raw_statistics_2q_nontrivial_nonentangled_state(forest):
     for sim_num in range(num_simulations):
         qc.qam.random_seed = sim_num+1
         expt_results = list(estimate_observables(qc, obs_expt, num_shots=num_shots,
-                                                 symmetrization_method=exhaustive_symmetrization))
+                                                 symm_type=-1))
         expt_results = list(calibrate_observable_estimates(qc, expt_results, noisy_program=p))
         raw_expectations.append([res.raw_expectation for res in expt_results])
         raw_std_errs.append([res.raw_std_err for res in expt_results])
@@ -1673,7 +1603,7 @@ def test_raw_statistics_2q_nontrivial_entangled_state(forest):
     for sim_num in range(num_simulations):
         qc.qam.random_seed = sim_num+1
         expt_results = list(estimate_observables(qc, obs_expt, num_shots=num_shots,
-                                                 symmetrization_method=exhaustive_symmetrization))
+                                                 symm_type=-1))
         expt_results = list(calibrate_observable_estimates(qc, expt_results, noisy_program=p))
         raw_expectations.append([res.raw_expectation for res in expt_results])
         raw_std_errs.append([res.raw_std_err for res in expt_results])
@@ -1871,7 +1801,7 @@ def test_bit_flip_state_fidelity_readout_error(forest):
         qc.qam.random_seed = sim_num+1
         expt_results = []
         results = estimate_observables(qc, process_exp,
-                                       symmetrization_method=exhaustive_symmetrization)
+                                       symm_type=-1)
         for res in calibrate_observable_estimates(qc, list(results), noisy_program=p):
             expt_results.append(res.expectation)
         expts.append(expt_results)
@@ -1907,7 +1837,7 @@ def test_dephasing_state_fidelity_readout_error(forest):
         qc.qam.random_seed = sim_num+1
         expt_results = []
         results = estimate_observables(qc, process_exp,
-                                       symmetrization_method=exhaustive_symmetrization)
+                                       symm_type=-1)
         for res in calibrate_observable_estimates(qc, list(results), noisy_program=p):
             expt_results.append(res.expectation)
         expts.append(expt_results)
@@ -1945,7 +1875,7 @@ def test_depolarizing_state_fidelity_readout_error(forest):
         qc.qam.random_seed = sim_num+1
         expt_results = []
         results = estimate_observables(qc, process_exp,
-                                       symmetrization_method=exhaustive_symmetrization)
+                                       symm_type=-1)
         for res in calibrate_observable_estimates(qc, list(results), noisy_program=p):
             expt_results.append(res.expectation)
         expts.append(expt_results)
@@ -1977,7 +1907,7 @@ def test_unitary_state_fidelity_readout_error(forest):
         qc.qam.random_seed = sim_num+1
         expt_results = []
         results = estimate_observables(qc, process_exp,
-                                       symmetrization_method=exhaustive_symmetrization)
+                                       symm_type=-1)
         for res in calibrate_observable_estimates(qc, list(results), noisy_program=p):
             expt_results.append(res.expectation)
         expts.append(expt_results)